Update import paths in npmBuild, npmInstall, readPackage, setVersion, writePackage, updatePackageRegistry, publishPackage, renamePackage, execWrapper, generationVersion, and aliasDependencies
