+++ .github/scripts/src/aliasDependencies.ts
@@ -1,8 +1,16 @@
 import * as fs from "fs";
-import { Package, depTypes, spacer } from "src";
+
+import { depTypes, spacer } from "src/main";
+import { Package } from "src/types/package";
 import { writePackage } from "./writePackage";
 
-export async function aliasDependencies(packagePath: string, packagesToRescope: string[], fromScope: string, toScope: string, dependencyVersion = "*") {
+export async function aliasDependencies(
+  packagePath: string,
+  packagesToRescope: string[],
+  fromScope: string,
+  toScope: string,
+  dependencyVersion = "*"
+) {
   spacer({ count: 40 });
   console.log(`Renaming dependencies in ${packagePath}`);
   const packageJson: Package = JSON.parse(fs.readFileSync(packagePath, "utf8"));
@@ -12,7 +20,10 @@ export async function aliasDependencies(packagePath: string, packagesToRescope: 
       for (const [dep, version] of Object.entries(deps)) {
         for (const pkg of packagesToRescope) {
           if (dep === pkg) {
-            const newVersion = `npm:${dep.replace(fromScope, toScope)}@${dependencyVersion}`;
+            const newVersion = `npm:${dep.replace(
+              fromScope,
+              toScope
+            )}@${dependencyVersion}`;
             console.log(`${depType}.${dep}: "${newVersion}"`);
             deps[dep] = newVersion;
             spacer({ count: 10 });
