+++ .github/scripts/src/execWrapper.ts
@@ -2,7 +2,11 @@ import * as exec from "@actions/exec";
 import * as fs from "fs";
 import path from "path";
 
-export async function execWrapper(command: string, args: string[], options: { cwd: string; }) {
+export async function execWrapper(
+  command: string,
+  args: string[],
+  options: { cwd: string }
+) {
   let cwd = options.cwd;
   if (fs.existsSync(cwd) && fs.lstatSync(cwd).isFile()) {
     cwd = path.dirname(cwd);
@@ -15,7 +19,7 @@ export async function execWrapper(command: string, args: string[], options: { cw
     },
     stderr: (data: Buffer) => {
       // console.error(data.toString());
-    }
+    },
   };
   await exec.exec(command, args, { cwd, listeners }).catch((err: any) => {
     console.error(err);
