+++ .github/scripts/src/publishPackage.ts
@@ -2,7 +2,11 @@ import path from "path";
 import { execWrapper } from "./execWrapper";
 import { updatePackageRegistry } from "./updatePackageRegistry";
 
-export async function publishPackage(cwd: string, registry: string, flags: string[] = []) {
+export async function publishPackage(
+  cwd: string,
+  registry: string,
+  flags: string[] = []
+) {
   console.log(`Publishing`, { cwd, registry, flags });
   const packageDir = path.dirname(cwd);
   console.log({ packageDir });
