+++ .github/scripts/src/readPackage.ts
@@ -1,6 +1,6 @@
 import * as fs from "fs";
-import { Package } from "src";
 
+import { Package } from "src/types/package";
 
 export function readPackage(packagePath: string): Package {
   return JSON.parse(fs.readFileSync(packagePath, "utf8"));
