+++ .github/scripts/src/types/package.ts
@@ -1,0 +1,10 @@
+export type Package = {
+  name?: string;
+  version?: string;
+  dependencies?: Record<string, string>;
+  devDependencies?: Record<string, string>;
+  peerDependencies?: Record<string, string>;
+  publishConfig?: {
+    registry?: string;
+  };
+};
