+++ .github/scripts/src/updatePackageRegistry.ts
@@ -1,7 +1,6 @@
 import { readPackage } from "./readPackage";
 import { writePackage } from "./writePackage";
 
-
 export function updatePackageRegistry(cwd: string, registry: string) {
   const packageJson = readPackage(cwd);
   packageJson.publishConfig = { registry };
